Avoids code duplication passing the number of time intcode should run
This commit adds the `times` paramter to the intcode function. It
provides a simple solution to make sure there are the number of `time`
values in the output array. For example when the problem requires 3
elements in the output, then run the intcode program 3 times.

Also, exposes some getters from the intcode class.